requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
cachetools==3.1.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Click==7.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.5
cssselect==1.0.3
dialogflow==0.5.2
Django==2.1.5
django-cors-headers==2.4.0
django-extensions==2.1.5
django-phonenumber-field==2.1.0
djangorestframework==3.9.1
elasticsearch==7.0.2
Flask==1.0.2
google-api-core==1.7.0
google-auth==1.6.3
googleapis-common-protos==1.5.8
gunicorn==19.9.0
grpcio==1.18.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10
lxml==4.3.1
MarkupSafe==1.1.0
nltk==3.4
numpy==1.16.2
pandas==0.24.2
parsel==1.5.1
phonenumbers==8.10.4
protobuf==3.6.1
psycopg2-binary==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pydialogflow-fulfillment==0.0.3
PyDispatcher==2.0.5
pydotplus==2.0.2
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
queuelib==1.5.0
requests==2.21.0
rsa==4.0
Scrapy==1.6.0
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.12.0
Twisted==18.9.0
unicode==2.7
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
w3lib==1.20.0
Werkzeug==0.15.2
zope.interface==4.6.0   
